Move tracking into folder
