+++ src/client/pages/OfferNew/index.tsx
@@ -15,7 +15,7 @@ import React from 'react'
 import { Redirect, useHistory, useRouteMatch } from 'react-router'
 import { useVariation, Variation } from 'utils/hooks/useVariation'
 import { useStorage } from 'utils/StorageContainer'
-import { getUtmParamsFromCookie, TrackAction } from 'utils/tracking'
+import { getUtmParamsFromCookie, TrackAction } from 'utils/tracking/tracking'
 import { Checkout } from './Checkout'
 import { FaqSection } from './FaqSection'
 import { Introduction } from './Introduction'
